Renamed a few files and reorganized some stuff.
