+++ public/src/css/default.css
@@ -23,116 +23,6 @@ h3 {
     font-size: 16px;
 }
 
-/** homepage */
-.homePage {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 100%;
-
-    background-color: rgb(30, 27, 26);
-}
-.mainContent {
-    width: 800px;
-    text-align: center;
-}
-
-.horizontalDivider {
-    margin: 12px 0;
-    width: 300px;
-    border: 1px solid rgb(242, 247, 244);
-}
-
-.socialButton {
-    width: 150px;
-    padding: 9px;
-    margin: 3px;
-
-    font-weight: 600;
-    background-color: rgb(71, 74, 114);
-    color: rgb(242, 247, 244);
-    border: 1px solid rgb(242, 247, 244);
-    border-radius: 2%;
-}
-
-.socialButton:hover {
-    background-color: rgb(62, 63, 105);
-    border: none;
-}
-
-.row {
-    display: flex;
-    flex-direction: row;
-}
-
-.column {
-    display: flex;
-    flex-direction: column;
-}
-
-/** template stuff */
 #app {
     height: 100%;
-}
-
-.page {
-    height: 100%;
-}
-
-.sidebar {
-    float: left;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: left;
-
-    width: 150px;
-    height: 100%;
-
-    background-color: rgb(23, 20, 19);
-}
-
-.sidebar-item {
-    padding: 18px 9px;
-}
-.sidebar-item:hover {
-    border: none;
-    background-color: rgb(71, 74, 114);
-}
-
-.contents {
-    height: 100%;
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: left;
-}
-
-.information {
-    display: flex;
-    flex-direction: column;
-    justify-content: start;
-    align-items: center;
-    flex: 1 1 400px;
-
-    background-color: rgb(30, 27, 26);
-}
-
-.widget {
-    display: flex;
-    justify-content: center;
-    flex: 1 1 400px;
-
-    background-color: rgb(45, 40, 39);
-}
-
-/** Homepage */
-.homeInformation {
-    margin: 0 24px;
-}
-
-/** For smaller screens */
-@media screen and (max-width:990px) {
-    .information {
-        flex-basis: 100%;
-    }
 }
