+++ public/src/js/Templates/page.css
@@ -1,0 +1,36 @@
+.page {
+    height: 100%;
+}
+
+.contents {
+    height: 100%;
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: left;
+}
+
+.information {
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+    align-items: center;
+    flex: 1 1 400px;
+
+    background-color: rgb(30, 27, 26);
+}
+
+.widget {
+    display: flex;
+    justify-content: center;
+    flex: 1 1 400px;
+
+    background-color: rgb(45, 40, 39);
+}
+
+
+/** For smaller screens */
+@media screen and (max-width:990px) {
+    .information {
+        flex-basis: 100%;
+    }
+}
