+++ public/src/js/Templates/sidebar.css
@@ -1,0 +1,20 @@
+.sidebar {
+    float: left;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: left;
+
+    width: 150px;
+    height: 100%;
+
+    background-color: rgb(23, 20, 19);
+}
+
+.sidebar-item {
+    padding: 18px 9px;
+}
+.sidebar-item:hover {
+    border: none;
+    background-color: rgb(71, 74, 114);
+}
